fix: simplify collection to nox exit-code filtering
